requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
eventlet==0.40.4
Flask==3.1.2
flask-cors==6.0.2
Flask-SocketIO==5.6.0
greenlet==3.3.1
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pymongo==4.16.0
python-dotenv==1.2.1
python-engineio==4.13.1
python-socketio==5.16.1
simple-websocket==1.1.0
Werkzeug==3.1.5
wsproto==1.3.2
